Refactor most of the tree stuff out of MainWindowController
